graph set up finish
